Update changelog, unit tests, and other small updates.
