+++ authorization_server/Dockerfile
@@ -1,5 +1,5 @@
 # 1. Pulling the dependencies
-FROM maven:3.8.5-openjdk-17 AS builder
+FROM maven:3.8.5-eclipse-temurin-17 AS builder
 # Change working directory in the container
 WORKDIR /opt/app
 # Copy all POMs first for better build caching
@@ -14,10 +14,10 @@ COPY ../resource_server/src ./resource_server/src
 COPY ../common_tools/src ./common_tools/src
 
 # compile code in /opt/app
-RUN mvn -B -e clean install -DskipTests
+RUN mvn -B -e clean install
 
 # 3. Preparing the runtime environment
-FROM openjdk:17-slim
+FROM eclipse-temurin:17
 
 WORKDIR /opt/app
