+++ pom.xml
@@ -87,14 +87,6 @@
 				<artifactId>spring-boot-maven-plugin</artifactId>
 				<version>3.3.0</version>
 			</plugin>
-			<plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-surefire-plugin</artifactId>
-                <version>3.0.0-M2</version>
-                <configuration>
-					<skipTests>true</skipTests>
-                </configuration>
-            </plugin>
 		</plugins>
 	</build>
