+++ resource_server/src/main/java/eu/europa/ec/eudi/signer/r3/resource_server/model/certificates/ejbca/EjbcaProperties.java
@@ -51,10 +51,6 @@ public class EjbcaProperties {
         this.countries = countries;
     }
 
-    public List<CountryConfig> getCountries() {
-        return countries;
-    }
-
     public String getCertificateAuthorityName(String country) {
         String def = null;
         boolean found = false;
